refactor: move VectorField type hint
