+++ src/galax/dynamics/integrate.py
@@ -1,14 +1,16 @@
 """:mod:`galax.dynamics.integrate`."""
 
-from ._src.integrate import core, funcs, utils
+from ._src.integrate import core, funcs, type_hints, utils
 from ._src.integrate.core import *  # noqa: F403
 from ._src.integrate.funcs import *  # noqa: F403
+from ._src.integrate.type_hints import *  # noqa: F403
 from ._src.integrate.utils import *  # noqa: F403
 
 __all__: list[str] = []
 __all__ += core.__all__
 __all__ += funcs.__all__
 __all__ += utils.__all__
+__all__ += type_hints.__all__
 
 # Cleanup
-del core, funcs, utils
+del core, funcs, utils, type_hints
